requirements.txt
accelerate==0.30.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bce-python-sdk==0.9.35
beautifulsoup4==4.13.4
bitsandbytes==0.43.1
blinker==1.9.0
blis==1.3.0
boto3==1.39.4
botocore==1.39.4
Brotli==1.1.0
bs4==0.0.2
build==0.10.0
cachetools==6.1.0
camelot-py==1.0.0
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.5
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dataclasses-json==0.6.7
datasets==2.19.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deskew==1.5.3
dill==0.3.8
distro==1.9.0
doclayout_yolo==0.0.4
docling @ git+https://github.com/docling-project/docling.git@95e70962f1d7cf1f339a88fde9c907111e194726
docling-core==2.42.0
docling-ibm-models==3.8.1
docling-parse==4.1.0
easyocr==1.7.2
einops==0.8.1
Elixir==0.7.1
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
extractable==1.0.2
faiss-cpu==1.11.0.post1
Faker==18.9.1
fast-langdetect==0.2.5
fastapi==0.116.1
fasttext-predict==0.9.2.4
ffmpy==0.6.0
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
Flask==3.1.1
flask-babel==4.0.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.3.1
ftfy==6.3.1
future==1.0.0
gensim==4.3.2
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.45
gpt4all==2.8.2
GPUtil==1.4.0
gradio==5.37.0
gradio_client==1.10.4
gradio_pdf==0.0.22
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
img2pdf==0.6.1
imgaug==0.4.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jpype1==1.6.0
json_repair==0.47.7
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langchain==0.1.14
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.3.27
langchain-text-splitters==0.0.2
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.1.147
language_data==1.3.0
latex2mathml==3.78.0
lazy_loader==0.4
lmdb==1.7.2
loguru==0.7.3
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.3.7
markdown-it-py==3.0.0
marko==2.1.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mineru==2.1.0
mkdocs==1.4.3
modelscope==1.28.0
more-itertools==10.7.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.48.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-nvjitlink-cu12==12.9.86
olefile==0.47
omegaconf==2.3.0
onnxruntime==1.22.1
openai==1.95.1
opencv-contrib-python==4.6.0.66
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
packaging==23.2
paddleocr==2.6.1.3
paddlepaddle-gpu==2.5.2
paddlex==3.1.2
pandas==2.3.1
pandas-stubs==2.0.2.230605
parso==0.8.4
pbr==6.1.1
pdf2docx==0.5.8
pdf2image==1.15.1
pdf2jpg==1.1
pdfminer.six==20250506
pdfplumber==0.11.7
pdftext==0.6.3
peft==0.10.0
pikepdf==9.10.2
pillow==11.3.0
pillow-avif-plugin==1.5.2
pillow_heif==1.0.0
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
poppler-utils==0.1.0
premailer==3.10.0
preshed==3.0.10
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.26.3
pyngrok==7.2.12
pyparsing==3.2.3
pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==8.4.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyyaml_env_tag==1.1
pyzmq==27.0.0
rapid-table==1.0.5
RapidFuzz==3.13.0
rapidocr-onnxruntime==1.4.4
rarfile==4.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
roboflow==1.2.1
robust-downloader==0.0.2
rpds-py==0.26.0
rtree==1.4.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.3
s3transfer==0.13.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
semchunk==2.2.2
sentence-transformers==5.0.0
sentencepiece==0.2.0
shapely==2.1.1
shellingham==1.5.4
simsimd==6.5.0
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.42
sqlalchemy-migrate==0.11.0
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.1
streamlit==1.47.0
stringzilla==3.12.5
svgwrite==1.4.3
sympy==1.13.1
tabula-py==2.10.0
tabulate==0.9.0
Tempita==0.6.0
tenacity==8.5.0
termcolor==3.1.0
thinc==8.3.6
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==0.9.16
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
toolz==0.12.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@6b3a1f2f5100a84e40138382a6955627b3b865ba
typer==0.16.0
types-pytz==2025.2.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.166
ultralytics-thop==2.0.14
unstructured==0.18.11
unstructured-client==0.40.0
urllib3==2.5.0
uv==0.7.21
uvicorn==0.35.0
visualdl==2.5.3
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
win32_setctime==1.2.0
wrapt==1.17.2
xlsxwriter==3.2.5
xlwt==0.7.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0